resolved test failures

+++ src/test/java/africa/nkwadoma/nkwadoma/infrastructure/adapters/output/persistence/loanManagement/LoaneePersistenceAdapterTest.java
@@ -220,10 +220,11 @@ class LoaneePersistenceAdapterTest {
 
     @AfterAll
     void cleanUp() throws MeedlException {
-        identityManagerOutputPort.deleteUser(userIdentity);
         loaneeRepository.deleteById(loaneeId);
         identityOutputPort.deleteUserById(userIdentity.getId());
-        loanBreakdownOutputPort.deleteAll(loanBreakdownList);
+        if (loanBreakdownList != null) {
+          loanBreakdownOutputPort.deleteAll(loanBreakdownList);
+        }
         loaneeLoanDetailsOutputPort.delete(loaneeLoanDetailId);
     }
 }
